requirements.txt
streamlit==1.22.0
opencv-python==4.7.0.72
numpy==1.24.3
pycryptodome==3.17
cryptography==41.0.2
pycryptodomex==3.17
